requirements.txt
flask==3.0.0
flask-restful==0.3.10
python-dotenv==1.0.1
llama-index-core==0.12.16
llama-index-embeddings-openai==0.3.1
pgvector==0.2.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.52.0
dashscope==1.14.0
